+++ src/program/mod.rs
@@ -7,7 +7,7 @@ use gl;
 use version::Api;
 use version::Version;
 
-pub use self::compute::ComputeShader;
+pub use self::compute::{ComputeShader, ComputeCommand};
 pub use self::program::Program;
 pub use self::reflection::{Uniform, UniformBlock, BlockLayout, OutputPrimitives};
 pub use self::reflection::{Attribute, TransformFeedbackVarying, TransformFeedbackBuffer, TransformFeedbackMode};